server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyYAML
python-box
psycopg2-binary
SQLAlchemy
python-dotenv
moviepy
opencv-python-headless
mediapipe==0.10.21
transformers
torch
spacy
ffmpeg-python
librosa          
soundfile
nltk
scipy        
scikit-learn      
joblib            
numpy             
pandas
openai-whisper
sentence-transformers
psycopg2-binary 